requirements.txt
langchain>=0.1.0
langchain-ollama>=0.0.1
langchain-community>=0.0.13
langchain-qdrant>=0.0.1
langchain-postgres>=0.0.1
python-dotenv>=1.0.0
pypdf>=3.17.1
qdrant-client>=1.7.0
flask>=2.0.0
psycopg2-binary>=2.9.0 